+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Less logging when wifi is off and we load from cache (#157)
 - Alpha: Add Provided & Secret Support (#152)
 - Alpha: x_datafile (#156)
+- Add single line action-controller output under rails.controller (#158)
 
 ## 1.2.1 - 2023-11-01
 - Update protobuf definitions (#154)